bot/requirements.txt
python-telegram-bot
pyotp